harpia: power_profile: fudge the radio on current draw
Radio time accouting changed on Oreo, causing radio active time
to be reported as too long. For some reason, the claimed radio active
current for harpia in the power_profile is nearly 1/4 that of lux,
which is probably wrong. Anyway, I'm reducing it to 50 to match the
fudged value on lux that seems to give reasonable results.

Change-Id: Ie44fc28615b9543fdafb92cd7df41f4862a1ef3f